move aggregator to separate namesace
